+++ src/mongocxx/lib/mongocxx/v1/pool.cpp
@@ -25,6 +25,7 @@
 #include <mongocxx/v1/auto_encryption_options.hh>
 #include <mongocxx/v1/client.hh>
 #include <mongocxx/v1/exception.hh>
+#include <mongocxx/v1/oidc_callback.hh>
 #include <mongocxx/v1/server_api.hh>
 #include <mongocxx/v1/tls.hh>
 #include <mongocxx/v1/uri.hh>
@@ -50,6 +51,7 @@ class pool::impl {
    public:
     mongoc_client_pool_t* _pool;
     v1::apm _apm;
+    v1::oidc_callback _oidc_callback;
 
     ~impl() {
         libmongoc::client_pool_destroy(_pool);
@@ -117,6 +119,10 @@ pool::pool(v1::uri const& uri, options opts) : pool{uri} {
 
     if (auto& opt = v1::client::options::internal::apm_opts(client_opts)) {
         internal::set_apm(*this, std::move(*opt));
+    }
+
+    if (auto& opt = v1::client::options::internal::oidc_callback(client_opts)) {
+        internal::set_oidc_callback(*this, std::move(*opt));
     }
 
     if (auto const& opt = v1::client::options::internal::auto_encryption_opts(client_opts)) {
@@ -406,6 +412,12 @@ void pool::internal::set_apm(pool& self, v1::apm v) {
     auto& _apm = impl::with(self)._apm;
     _apm = std::move(v);
     v1::apm::internal::set_apm_callbacks(impl::with(self)._pool, _apm);
+}
+
+void pool::internal::set_oidc_callback(pool& self, v1::oidc_callback v) {
+    auto& _oidc_callback = impl::with(self)._oidc_callback;
+    _oidc_callback = std::move(v);
+    v1::set_oidc_callback(impl::with(self)._pool, _oidc_callback);
 }
 
 v1::client::options& pool::options::internal::client_opts(options& self) {