+++ src/mongocxx/lib/mongocxx/v_noabi/mongocxx/pool.cpp
@@ -59,6 +59,10 @@ pool::pool(v_noabi::uri const& mongodb_uri, v_noabi::options::pool const& option
         v1::pool::internal::set_apm(_pool, v_noabi::to_v1(*opt));
     }
 
+    if (auto const& opt = client_opts.oidc_callback()) {
+        v1::pool::internal::set_oidc_callback(_pool, *opt);
+    }
+
     if (auto const& opt = client_opts.auto_encryption_opts()) {
         bson_error_t error = {};