+++ src/mongocxx/test/v_noabi/client_helpers.cpp
@@ -54,10 +54,21 @@ using bsoncxx::builder::basic::kvp;
 using bsoncxx::builder::basic::make_document;
 
 namespace {
+
+mongocxx::uri test_uri() {
+    if (auto const* oidc_user = std::getenv("OIDC_ADMIN_USER")) {
+        auto const* oidc_pwd = std::getenv("OIDC_ADMIN_PWD");
+        REQUIRE(oidc_pwd);
+        // The OIDC test server requires auth. For test setup, use username/password.
+        return mongocxx::uri{"mongodb://" + std::string(oidc_user) + ":" + std::string(oidc_pwd) + "@localhost:27017"};
+    }
+    return mongocxx::uri{"mongodb://localhost:27017"};
+}
+
 // These frequently used network calls are cached to avoid bottlenecks during tests.
 document::value get_is_master() {
     static auto reply = []() {
-        auto client = mongocxx::client{mongocxx::uri{}, test_util::add_test_server_api()};
+        auto client = mongocxx::client{test_uri(), test_util::add_test_server_api()};
         return client["admin"].run_command(make_document(kvp("isMaster", 1)));
     }();
     return reply;
@@ -65,7 +76,7 @@ document::value get_is_master() {
 
 document::value get_server_status() {
     static auto status = []() {
-        auto client = mongocxx::client{mongocxx::uri{}, test_util::add_test_server_api()};
+        auto client = mongocxx::client{test_uri(), test_util::add_test_server_api()};
         return client["admin"].run_command(make_document(kvp("serverStatus", 1)));
     }();
     return status;
@@ -73,7 +84,7 @@ document::value get_server_status() {
 
 bsoncxx::stdx::optional<document::value> get_shards() {
     static auto shards = []() {
-        auto client = mongocxx::client{mongocxx::uri{}, test_util::add_test_server_api()};
+        auto client = mongocxx::client{test_uri(), test_util::add_test_server_api()};
         return client["config"]["shards"].find_one({});
     }();
     return (shards) ? shards.value() : bsoncxx::stdx::optional<document::value>{};
@@ -280,7 +291,7 @@ std::string get_server_version() {
 document::value get_server_params() {
     // Cache reply.
     static auto reply = []() {
-        auto client = mongocxx::client{mongocxx::uri{}, test_util::add_test_server_api()};
+        auto client = mongocxx::client{test_uri(), test_util::add_test_server_api()};
         return client["admin"].run_command(make_document(kvp("getParameter", "*")));
     }();
 
@@ -590,7 +601,7 @@ cseeos_result client_side_encryption_enabled_or_skip_impl() {
 
 #endif // defined(MONGOC_ENABLE_CLIENT_SIDE_ENCRYPTION)
 
-std::string getenv_or_fail(const std::string env_name) {
+std::string getenv_or_fail(std::string const& env_name) {
     auto val = std::getenv(env_name.c_str());
     if (!val) {
         FAIL("Please set the environment variable: " << env_name);