+++ src/mongocxx/test/v_noabi/client_helpers.hh
@@ -242,7 +242,7 @@ cseeos_result client_side_encryption_enabled_or_skip_impl();
 #define CLIENT_SIDE_ENCRYPTION_ENABLED_OR_SKIP() SKIP("linked libmongoc does not support client side encryption")
 #endif // defined(MONGOC_ENABLE_CLIENT_SIDE_ENCRYPTION)
 
-std::string getenv_or_fail(const std::string env_name);
+std::string getenv_or_fail(std::string const& env_name);
 
 } // namespace test_util
 } // namespace mongocxx